[Refactor] Re-enable gdb attach on crash.
PID struct seems pretty self-contained so let's move it to
core.h/core.cpp in taichi/common/. This ensures no dependency from
taichi/common to taichi/system. This PR also re-enables the developer
feature that automatically attaching when crash happens.

ghstack-source-id: 3aa603678300ccb529b3ae86ffe3a5b377d140ce
Pull Request resolved: https://github.com/taichi-dev/taichi/pull/2682
